Fix more and more bugs

+++ backend/webapp/src/main/java/ar/edu/itba/paw/webapp/dto/TicketDto.java
@@ -52,7 +52,7 @@ public class TicketDto {
         return ticketId;
     }
 
-    public void setTicketId(int ticketId) {
+    public void setTicketId(long ticketId) {
         this.ticketId = ticketId;
     }
